create image processor module
