+++ app/processors/geo_concerns/processors/base_geo_processor.rb
@@ -1,5 +1,3 @@
-require 'mini_magick'
-
 module GeoConcerns
   module Processors
     module BaseGeoProcessor
@@ -35,28 +33,6 @@ module GeoConcerns
         def self.temp_path(path)
           time = (Time.now.to_f * 1000).to_i
           "#{File.dirname(path)}/#{File.basename(path, File.extname(path))}_#{time}"
-        end
-
-        ## TODO: Move to new image processor
-        #
-        # Uses imagemagick to resize an image and convert it to the output format.
-        # Keeps the aspect ratio of the original image and adds padding to
-        # to the output image. The file extension is the output format.
-        # @param in_path [String] file input path
-        # @param out_path [String] processor output file path.
-        # @param options [Hash] creation options
-        # @option options [String] `:output_size` as "w h" or "wxh"
-        def self.convert(in_path, out_path, options)
-          image = MiniMagick::Image.open(in_path) # copies image
-          image.combine_options do |i|
-            size = options[:output_size].tr(' ', 'x')
-            i.resize size
-            i.background 'white'
-            i.gravity 'center'
-            i.extent size
-          end
-          image.format File.extname(out_path).gsub(/^\./, '')
-          image.write(out_path)
         end
       end
